server/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-decouple
pytest
pytest-django
djangorestframework-simplejwt
django-cors-headers
numpy
pandas
scikit-learn
nltk
lightfm>=1.17
joblib
